Make md resync thread pause when freezer is on.
Make the md resync thread pause when the freezer is on, so that resyncs
don't interfere with the process of hibernating and resuming. Ideally, the
resync thread wouldn't start until after we've checked whether to resume. I
need to talk to the md guys about this, but this workaround should help for
the meantime.

Signed-off-by: Nigel Cunningham <nigel@suspend2.net>