Fix failure patch in mark_pages_for_pageset2.
The failure path for allocating struct attention_list items was wrong. We
want to not go ahead with marking the tasks as pageset1, and just skip
through to freeing any memory that had been allocated. Without this change,
we'll oops in the next step (recording the tasks that need attention),
because we just assume there that there are enough structs allocated.

Signed-off-by: Nigel Cunningham <nigel@suspend2.net>