Add missing freezer_state export.
Adding the freezer_is_on() test to the md resync thread broke building that
code as a module, because I didn't export the freezer_state symbol. Address
that by adding an EXPORT_SYMBOL. I'm explicitly deciding to make this one
not EXPORT_SYMBOL_GPL so it can work in all contexts.

Signed-off-by: Nigel Cunningham <nigel@suspend2.net>