[Suspend2] Play catchup with powerdown code & simplify.
Simplify the powerdown code in preparation for the hibernation ops, added
with 2.6.22-rc1. Also play catchup, adding suspend_console and
resume_console calls that I didn't notice came in with 2.6.19. Finally,
reorder the pm_ops->prepare in line with the latest wisdom on the order in
which things should be done.

Signed-off-by: Nigel Cunningham <nigel@suspend2.net>