Revert papering over of __init/CPU_HOTPLUG issues.
The modification this patch reverts was originally added because of issues
with code apparently marked __init being called from cpu hotplugging. I
haven't yet found the time to find the culprit, and it may have been fixed
upstream in the meantime. The issue also only appeared (at least on my
machine, IIRC) with DEBUG_PAGEALLOC. I'll therefore reverse this patch for
now, in the hopes of a real solution being found.

Signed-off-by: Nigel Cunningham <nigel@suspend2.net>