Improve support for controlling timeout for early messages.
When we show messages very early in the process, we have previously waited
up to 25 seconds before doing the default option. This patch implements a
new suspend2_wait kernel parameter which can control how long to wait, and
a kconfig option to set the default. A value of -1 means wait forever. 0
means don't wait at all and positive values indicate a number of seconds to
wait. Default behaviour is to preserve the previous 25 second delay.

Signed-off-by: Nigel Cunningham <nigel@suspend2.net>