Allow modules to put files in /sys/power/suspend2.
Add code to interpret "[ROOT]" as a special directory name, meaning that
a module's files live in /sys/power/suspend2, rather than a subdirectory.
This will be used for the conversion of the powerdown file to a module in
its own right, so that its files can still live in the suspend2 directory
for backwards compatibility.

Signed-off-by: Nigel Cunningham <nigel@suspend2.net>