Address issues with preparing for & cleaning up after atomic copy/restore.
Fix some typos introduced by the recent addition of the go_atomic &
end_atomic routines. In addition to this and as part of those changes, I
applied the result of a discussion on the linux-pm list about what order
things should be done in. This seems to have introduced problems, so I'll
revert the changes, at least temporarily.

Signed-off-by: Nigel Cunningham <nigel@suspend2.net>