Fix a race where pressing escape could lead to fs corruption.
Fix a race where pressing escape after the atomic restore (at resume time)
but prior to setting the NOW_RESUMING flag would cause TuxOnIce to act
like we're aborting at suspend time instead, and only reread the portion
of pageset2 that would have been overwritten by an atomic copy of pageset1.

The fix is simple - set the flag as soon as the atomic restore is done, and
move the check for aborting in the highlevel code to after the check for
resume time.

Signed-off-by: Nigel Cunningham <nigel@suspend2.net>