Handle large amounts of swapspace better.
Jonas Petersson reported an issue where, having approximately 4GB of swap,
he was unable to hibernate. Calculations of the amount of storage available
were being messed up because we (rightly) remove an allowance for the
headers written for each page (assuming no compression). The math
involved, however, exceeded the size of an int on 32 bit architectures.
To work around this, we now first calculate the allowance, then subtract it,
rather than doing it in one step:

a * PAGE_SIZE / (PAGE_SIZE + allowance) =
a - ((a * allowance) / (PAGE_SIZE + allowance))

Where allowance < PAGE_SIZE, we use less bits in the calculation. On 32bit,
we're using around 8 fewer bits, meaning swap can then be 256 times as large
(1T) before we run into the same issue.

Signed-off-by: Nigel Cunningham <nigel@tuxonice.net>