[TuxOnIce] Add support for transferring boot kernel data across versions.
Add support for transferring data from the boot kernel to the resumed
kernel. This is achieved by allocating a page for receiving the data
when hibernating, and storing its address in the image header.

At resume time, the boot kernel allocates a non-conflicting page (there
is no guarantee that its boot-kernel-data struct won't be overwritten
during the atomic restore), creates an extra pbe on the end of the
lowmem list (mapping this page to the one the resuming kernel allocated)
and copies its boot kernel data to the non-conflicting page. All of this
puts the boot kernel's struct in the resumed kernel's allocated page
after the atomic restore is done.

The resumed kernel then looks at the size and version of the boot kernel's
data. The initial implementation ignores the version number and copies
back a number of bytes which are the minimum of its own struct size and
the size of the boot kernel's struct. This implies that later versions
may append new fields, but not remove earlier ones.

Signed-off-by: Nigel Cunningham <nigel@tuxonice.net>