[TuxOnIce] Rework max-outstanding-io logic.
Rework the logic for limiting the amount of outstanding i/o so that
when writing, we submit I/O as quickly as CPU speed and available
memory allow. Allow the user to set a 'target' amount of I/Os, though,
so that in low memory situations we can seek to ensure we can have at
least that many bios on the go at once.

I still need to find an effect way of ensuring we don't run out of
memory completely - we want to leave some free so that the non-TuxOnIce
code can still do allocations. Ideally, this would involve determining
whether we're close to the page allocation watermarks.

Signed-off-by: Nigel Cunningham <nigel@tuxonice.net>