[TuxOnIce] Implement minimum queue length limit when writing bios.
Yes, this essentially undoes the removal of submitting in batches
that was done pre-rc3.

When writing an image, we fill pages with the data to be written, and
allocate a new page when the current one is filled. Full pages are added
to a queue, and processed after we've dropped the page-filling mutex
so that other threads can do their work. Prior to this patch, we were
always immediately starting to flush the queue (unless another thread
is already doing it). This patch adds a user-tunable minimum queue
length before we start flushing (unless we're finishing writing the
section of the image). This gives a modest performance improvement
on my SMP desktop. Writing goes from 99MB/s (with the trigger set to 1)
to 102MB/s (default of 25).

Signed-off-by: Nigel Cunningham <nigel@tuxonice.net>