[TuxOnIce] Add support for running commands pre and post hibernate.
Add support for running a usermode helper before starting to hibernate
and after resuming. Default values may be configured at compile time to
facilitate the use of TuxOnIce + hibernate script as a drop-in replacement
for swsusp. TuxOnIce does not continue running until the script completes.
At the moment, no notice is taken of the command's return code. I hope
to change this in the near future.

Signed-off-by: Nigel Cunningham <nigel@tuxonice.net>