[TuxOnIce] Throttle I/O submission when low on memory.
Add support for throttling I/O submission when we get low on memory.
On the first failure to allocate memory, find out how much memory is
currently free (this will be > 0 due to watermarks) add 50 and set that
as the point at which we'll wait.

The throttle is applied in get_io_info_struct. Each time we enter that
routine, we check if the number of free pages is less then the throttle
value, and wait for I/O completion & cleanup until enough memory is again
free.

We could potentially fall into a situation where there is no I/O in
progress but we don't have enough memory free for the throttle criteria.
This shouldn't happen because image preparation should free enough memory
prior to starting I/O. Nevertheless, we check there is some outstanding
I/O so as to avoid looping eternally here.

Signed-off-by: Nigel Cunningham <nigel@tuxonice.net>