[TuxOnIce] No default pre & post-hibernate commands for now.
I was going to have default pre and post-hibernate commands, calling
the hibernate script with --pre-hibernate and --post-hibernate
respectively. On further reflection, however, I think this is a bad idea.
For people who don't have up-to-date hibernate scripts (ones that will
understand the new args), it will probably result in the script returning
an error code, which will in turn cause the kernel to abort. So, for now
at least, I'll leave the parameter blank by default.

Signed-off-by: Nigel Cunningham <nigel@tuxonice.net>