[TuxOnIce] Disable lockdep while we're looking at whether to resume.
Turn lockdep off while considering whether to resume. Lockdep complains
when we try to use name_to_dev_t in the swap & file allocators because
sysfs buffer mutexs were taken when we initiated resuming, and will be
taken again when name_to_dev_t runs. Disable lockdep here so we don't
have to mess with code that's not ours.

Signed-off-by: Nigel Cunningham <nigel@tuxonice.net>