[TuxOnIce] Improve progress bar updating.
Improve updating of the progress bar by using atomic_sub_return rather than
atomic_dec and a later read (at a point when the value could have changed).
We then don't have to worry about which thread sees the point at which we
should be updating progress.

As part of this, also fix the initial value of io_nextupdate to be base + 1
rather than 0.

Signed-off-by: Nigel Cunningham <nigel@tuxonice.net>