[TuxOnIce] Make i/o limits tunable again.
In the course of the latest set of patches, I've made the tunable we did
have (target_outstanding_io) useless, and added new variables to control
limits. Now that things are settling, let's address this, making
target_outstanding_io useful again. It now sets an upper limit to the
number of pages we have submitted for read or write at a time.

Signed-off-by: Nigel Cunningham <nigel@tuxonice.net>