[TuxOnIce] Make default target_outstanding_io 1024.
When I was only using the text ui, 256 seemed like a reasonable default.
Having used the fbsplash one for a while, though, 1024 seems better. I
guess the fbsplash code is slow enough such that all of the I/O in progress
when we start a progress update completes by the time we have finished the
progress update. We'd therefore be doing bursts of I/O between progress
updates, which would account for the lower throughput.

Signed-off-by: Nigel Cunningham <nigel@tuxonice.net>