Properly cleanup swap allocator when not resuming.
Properly cleanup when we read the image header, but then don't go
on to resume (perhaps because we were just looking to see if there
is an image).

This fixes the problem where extents spill over into a second page
(or more) of the header, and we can't then resume.

Signed-off-by: Nigel Cunningham <nigel@tuxonice.net>