[TuxOnIce] Don't spit the dummy immediately if can't submit readahead.
Don't immediately error out if we can't submit readahead. Instead, stop
trying to submit new readahead and get on with using what we already have.
We'll try again if we run out of data.

While we're at it, don't be silent about the cause if we return -EIO for
some other reason.

Signed-off-by: Nigel Cunningham <nigel@tuxonice.net>