[TuxOnIce] Fix unbalanced _toi_try_resume error path.
Fix failure to called toi_finish_anything if initialisation of modules
fails. This would cause the atomic_t actions_running to not be
decremented, stopping subsequent attempts at hibernating. It would also
leave the process running with the kernel segment, not release
tuxonice module references, and leave the process bound to cpu0. Urgh!

Signed-off-by: Nigel Cunningham <nigel@tuxonice.net>