[TuxOnIce] Use binary signature for swap allocator.
Use the new binary signature for the swap allocator. To achieve this,
we take the information pointing to the first block of the header and
store it at the start of the swap header page, and backup the contents
of those bytes at the start of the first image header page.

Signed-off-by: Nigel Cunningham <nigel@tuxonice.net>