[TuxOnIce] Fix extra pages allowance autotuning.
At some stage in the past (apparently before I started using Git), I
managed to break autotuning of the extra pages allowance by clamping
the sysfs setting to MIN_EXTRA_PAGES_ALLOWANCE. Revert this. The idea
was right (if you're going to set a minimum, it should be enforced).
Given, however that we want to allow autotuning and there's nothing
technically wrong with having an extra pages allowance of 1...

Signed-off-by: Nigel Cunningham <nigel@tuxonice.net>