[TuxOnIce] Apply header reservation when size updated too.
I made a mistake in thinking I could safely apply the
header reservation in the routine that fills the list of
blocks. If swap requirements don't increase, that routine
is only called once. If header requirements increase,
we can end up with a smaller header size still applied,
resulting in the bio code catching the issue when
writing the header. Re-apply the reservation if it changes,
as well as in the first case.

Signed-off-by: Nigel Cunningham <nigel@tuxonice.net>