[TuxOnIce] Fix problem with using !LZF for compressor.
On resume, we were preparing the LZF compressor at initialisation time.
That's fine if we're using LZF, but if not, we weren't cleaning it up
and initialising the right compressor if something else is being used.
To fix that, check when loading configuration info whether the compressor
is different and if so, do the cleanup & reinit.

Signed-off-by: Nigel Cunningham <nigel@tuxonice.net>