[TuxOnIce] Ensure userui seeks working even if userspace dies.
Prior to this patch, we have a problem if userspace unexpectedly dies
without acknowledging our CLEANUP message. We will not be able to
run the helper again until after a reboot. This patch addresses that
issue by calling the cleanup routines that would have been run in
response to userspace sending its ACK. They are run when starting
another cycle, or when unloading the userui or usm modules (the
cleanup routine correctly handles being called when there is no cleanup
to do).

Signed-off-by: Nigel Cunningham <nigel@tuxonice.net>