[TuxOnIce] Don't do full prep/cleanup for reading simple sysfs entries.
Whoopie had a problem wherein userui would get unexpectedly killed.
We traced the issue to the recent modifications to start/finish_anything.
Since userui currently reads some of our sysfs entries, it seems that
the switch to a simple mutex in start_anything was causing it to stop
(waiting for the mutex) and get killed by us trying to freeze it.

The real fix is for userui to get all it's info via netlink, but for
now, I'll make the sysfs read routine not do a full setup/cleanup for
simple value reads.

Whoopie has confirmed that this fixes his issue.

Signed-off-by: Nigel Cunningham <nigel@tuxonice.net>