[TuxOnIce] Reinstate late_initcall of software_resume.
Reinstate the late initcall of software_resume. Without it, people without an
initrd/initramfs can't resume.

There is, however, a downside (which is why it was removed in the first place):
with the late_initcall reinstated, resuming will start before the initrd/ramfs
runs if that's possible. This means that userui won't run until after we've
switched back to the old kernel. For this reason, I will also add a means of
controlling the initcall by default (compile time) and from the boot command line.

Signed-off-by: Nigel Cunningham <nigel@tuxonice.net>