[TuxOnIce] Add option to ignore the late_initcall of software_resume.
The late_initcall of software resume was removed for a long time,
but that's not really the right solution. It breaks resuming for
people who don't use initramfses or initrds. A few commits ago, the
late initcall was reinistated. This commit completes fixing the
issue, by adding a compile time option which lets the user choose
whether to ignore the initcall. This compile time option can be
overridden by the command line option toi_initramfs_resume_only.

With this in place, we can support both resuming without using
an initramfs/initrd AND resuming using an initrd/ramfs, with
the code built in and a config that could allow us to resume at
the late_initcall point. (This would be wanted where you have
userui in your initramfs/initrd and want it to run for the first
part of resuming).

Signed-off-by: Nigel Cunningham <nigel@tuxonice.net>