[TuxOnIce] Complete support for monitoring outstanding I/O.
With the initial work on monitoring outstanding I/O, we only
monitor writes that have already been submitted. This ignores
pages that have been queued, but not yet submitted because of
a lack of memory. This patch adds recording of how many pages
are in the queue, which in turn enables us to monitor the total
number of pages that are waiting to be written.

Rather than updating the output in steps of 10MB, we also switch
to doing 5 updates between the initial value and the completion
of all I/O (ie every 20%).

We also separate the monitoring from the finish_all_io code, and
call this instead of finish_all_io from the end of the I/O
submission loop if we're multithreaded and the first thread to
stop submitting I/O. The last worker now needs to call
finish_all_io, in case the thread on CPU0 is doing the monitoring
(flushing of the write queue will otherwise not occur).

Signed-off-by: Nigel Cunningham <nigel@tuxonice.net>