[TuxOnIce] No pageset2 if it's not needed.
We don't always need to save the image in two parts. Add optional
functionality which checks whether a second pageset is needed, and
stores everything in one pageset if the user has enable the option and
it's possible to do so. I plan on making this the default, but will give
it some more testing first.

Signed-off-by: Nigel Cunningham <nigel@tuxonice.net>