Throttle I/O according to disk throughput.
Throttle the amount of I/O we submit according to our measurements of
the rate at which raw pages that are being written/read, as well as when
we run low on memory. Do this by having one of the I/O threads call a
function in the allocator (if it exists) every 1/10th of a second or so,
telling the function how many 10ths of a second have elapsed since we
started submitting I/O. The allocator (or, in the case of swap & file
support, the block I/O code) records how many I/O pages have been
completed to date, and uses this to update the level at which we
throttle. The throttle is reset at the start of each pageset that is
read or written, and is initially calculated as the number of pages
being completed every tenth of a second.

Prior to this modification, on laptops where the CPU is significantly
faster than the disk throughput, we would have a long wait at the end of
compressing pages, while the disk caught up. With this patch, that wait
will be in the region of 1/10th of a second worth of I/O, and the wait
will be spread through the whole period in which the pageset is written.
This in turn allows us to run animations in userui more smoothly.

While we're at it, cleanup the 'reasons we waited on I/O' array - some
reasons are no longer used and get rid of the TARGET_OUTSTANDING_IO
macro.

Signed-off-by: Nigel Cunningham <nigel@tuxonice.net>