Tweak I/O throttling algorithm.
Update the throttle value each half a second rather than tenth of a
second, giving more time for us to get a decent average before we start
applying the throttle. We continue to limit the throttle to one tenth of
a second worth of throughput.

Keep updating the throttle every 1/2s to allow for different throughput
in different parts of the storage medium.

Signed-off-by: Nigel Cunningham <nigel@tuxonice.net>