Improve handling of empty readahead queue.
On SMP, we can potentially go to get a page from the readahead queue and
find that the list is empty. If we are the thread that's submitting new
readahead, we should do this immediately, rather than waiting forever
for someone else to do it. In addition, we can use a wait queue if we're
one of the other threads instead of polling.

Signed-off-by: Nigel Cunningham <nigel@tuxonice.net>