Call bio_queue_flush_pages from throttle and handle errors.
If we're over the level where we want to throttle I/O, we should make
sure we've submitted all the available pages (there's a possible
deadlock if we wait for I/O that hasn't been submitted yet). In
addition, we need to use the result of submitting the I/O, passing any
errors back up through the call chain.

Signed-off-by: Nigel Cunningham <nigel@tuxonice.net>