Report shrink_all_memory results.
Report the results of attempts at freeing memory. Since
shrink_all_memory rarely frees as much memory as is requested, we should
have this, so as to be able to point the finger if hibernating fails.

I should really try to improve shrink_all_memory, but that's a patch
series on its own, I expect.

Signed-off-by: Nigel Cunningham <nigel@tuxonice.net>