Fix UP hangs when reading caches.
A number of people reported hangs with the 20090206 current patches,
which, it turns out, are related to running single threaded. We can't
wait for memory to come free when the thread doing the waiting is the
only one that's going to ever get around to using and freeing the pages
that have already been submitted. The solution is to drop out of
submitting readahead instead of waiting, and not treat that situation as
an error (get on with using the pages that have been submitted already).

We also need to not even bother to check the throttle in the submit
function if we're reading pages - we're already committed to submitting
the I/O. In the case of writing, we can check the throttle and wait for
previously submitted I/O to complete and associated pages to be freed.

Signed-off-by: Nigel Cunningham <nigel@tuxonice.net>