Fix do_rw_loop return value.
When pressing escape to cancel a cycle at the "Powering down now"
message, I discovered that the system would oops after rereading
pageset2. This was because do_rw_loop returns non-zero if the aborted
flag is set, rather than iff we failed to do all of the I/O that should
have been done. Fix this so that do_rw_loop only returns non-zero if the
I/O it should have done wasn't done.

Signed-off-by: Nigel Cunningham <nigel@tuxonice.net>