hw_detection: Initial implementation
Impact: Abort resume if a dmi mismatch is detected

Some situations are not resume safe, for instance when the hard drive is
put in another machine. In that case, we shouldn't try to resume but
rather invalidate the image and do a full boot.

To perform this detection, we rely on the DMI_PRODUCT_UUID that is
stored in the hibernate image.

Testing Done: hibernated on a T500. Put the drive on a DELL E6400.
Tried to resume. The kernel printed the mismatch message and did a full
boot. Hibernated on the DELL. Put the drive back in the T500. Tried to
resume. Same scenario: The kernel printed the mismatch message and did a
full boot.

Signed-off-by: Pierre-Alexandre Meyer <pierre@mouraf.org>