Don't panic when we aborted doing I/O.
The logic in do_rw_loop needs a little more tweaking after the recent
modifications to allow us to return a more useful result code. We
shouldn't check that all I/O was done if the user chose to abort. If we
reach this point, we're writing the image rather than reading it, and so
will want to back out.

Signed-off-by: Nigel Cunningham <nigel@tuxonice.net>