[TuxOnIce] Don't assume the swap signature in memory is up to date.
The swap signature that we have in memory can be out of date if it was
retrieved prior to an atomic copy and the completion of writing an
image. Prior to this patch, this would wrongly prevent us from resetting
the signature.

Observed when testing the extra-pages-allowance-too-small code, but
others have reported the issue before now (I just haven't been able to
see what caused the issue).

Signed-off-by: Nigel Cunningham <nigel@tuxonice.net>