[TuxOnIce] Add support for compcache.
Add support for compressed, ram based swap. Following the comments in
the compcache code, we check for devices where the name of the disk
begins with 'ram'. While we allocate swap from such devices, that's all
we do - we ignore them completely when it comes to building our chains
of sectors that will be used for reading and writing the image.
Allocated swap on these devices is not immediately freed, to help with
reliability in low memory situations (it will make the vm free memory
instead of trying to 'swap' further pages to the compcache device).

Signed-off-by: Nigel Cunningham <nigel@tuxonice.net>