[TuxOnIce] Stop potential OOM while getting pageset1 load addresses.
Milan reported in bug 413 that he was getting an OOM when beginning to
resume. This was caused by the recent change to freezing processes
asynchronously, combined with the fact that we allocate all available
memory when finding pages into which to locate pageset1.

Rather than switch back to synchronously freezing processes, this patch
disables non-boot cpus and irqs while we're getting the pageset1 load
addresses, to avoid the OOM. I can see that another potential solution
would be to combine allocating the pages with seeing whether and how we
can use them. That would be racy and perhaps unreliable with larger
pageset 1 sizes, so I'll probably stick with this.

Signed-off-by: Nigel Cunningham <nigel@tuxonice.net>