[TuxOnIce] Work on fixing occasional no more readhead issue.
I've had a couple of reports of TuxOnIce sometimes failing to read
pageset1 properly, BUGing in the !more_readahead test in
bio_get_next_page_read. This patch improves the more_readahead logic. It
shifts all the calculation of whether we have more pages to read to
go_next_page, putting it in the right place (we call go_next_page, then
immediately read/write that page, so positions saved are always the
last page in the previous section rather than the first page in the
section number). Finally, returning -ENODATA to start_new_readahead
shouldn't propogate to callers. I'm not sure I fully understand how we
can to hit the BUG(), but perhaps these fixes will stop us getting it in
future.

Signed-off-by: Nigel Cunningham <nigel@tuxonice.net>