[TuxOnIce] Enable LZO to work properly.
Thanks to Andreas Robinson, I now understand why LZO wasn't working
properly. The issue is that it requires an output buffer larger than
PAGE_SIZE, because (unlike LZF) it has a worst case result that is
larger than PAGE_SIZE. Cryptoapi doesn't currently allow us to query
exactly how big an output buffer needs to be and LZO doesn't check that
you've given it a buffer large enough. To work around this (and allow
for other algorithms that might have the same issue with a larger output
buffer requirement), vmalloc an output buffer 2 * PAGE_SIZE and use it
for compressing. As before, however, only use the output if it is <
PAGE_SIZE. If output is >= PAGE_SIZE, we simply store the input buffer.

This makes LZO work fine. I did also try deflate and it still didn't
work, so I'll do more work on discovering what's going on there and
prepare a separate patch (if necessary) for it.

Signed-off-by: Nigel Cunningham <nigel@tuxonice.net>