[TuxOnIce] Fix header reservation and debugging.
For a long time, reserving space for the header has been somewhat
broken, but I didn't know it. The brokenness was mostly hidden by a bug
which caused us to reserve one page more than is needed, a bug which was
fixed in a post-3.0 commit. Now that the brokenness has been properly
exposed, I've been able to properly fix it.

This patch addresses the issue by:
- Making debugging messages show up in the logs even if you're not using
  userui (KERN_DEBUG instead of KERN_INFO);
- Displaying all relevant details if we detect there is insufficient
  space (because allocation and usage have gotten out of sync again);
- Enabling us to back out of hibernation if we discover that header
  space is insufficient.
- Apply the header reservation without needing the hack to avoid the
  extent state eof message.
- Just set the header reservation size in allocators' reserve header
  space routines and apply it in the allocate_storage routines - even if
  there are no new pages to be allocated!

Signed-off-by: Nigel Cunningham <nigel@tuxonice.net>